テストのエラーを対処

+++ tests/c_backend.rs
@@ -140,7 +140,7 @@ fn test_render_for_loop() {
         vec![],
     );
     let expected = r#"
-for (int i = 0; i < N; i++) { }
+for (size_t i = 0; i < N; i++) { }
 "#;
     assert_render(ast, expected);
 }
